requirements.txt
fastapi[standard] == 0.115.11
uvicorn[standard] == 0.34.0
pydantic == 2.10.6
python-dotenv == 1.0.1

sqlalchemy == 2.0.39
psycopg2-binary == 2.9.10

python-jose[cryptography] == 3.4.0
passlib[bcrypt] == 1.7.4

aiohttp==3.8.6
requests==2.31.0
python-multipart==0.0.20
httpx==0.25.1
logging-plus==1.2.0
tenacity==8.2.3
pytest==7.4.3
pytest-asyncio==0.21.1

pyjwt == 2.10.1